add percentage percision

+++ scripts/utils.ts
@@ -29,7 +29,7 @@ export const encodeRoundParameters = (params: any[]): string => {
       "tuple(uint256 applicationsStartTime, uint256 applicationsEndTime, uint256 roundStartTime, uint256 roundEndTime)",
       "uint256",
       "address",
-      "uint8",
+      "uint32",
       "address",
       "tuple(tuple(uint256 protocol, string pointer), tuple(uint256 protocol, string pointer))",
       "tuple(address[] adminRoles, address[] roundOperators)"
